Show project with anchor when no commit hash passed
